Refactor tests˚
