+++ brave-kafka/src/test/java/gardentest/ConsumersTest.java
@@ -3,7 +3,8 @@ package gardentest;
 import brave.extension.KafkaExtension;
 import brave.extension.KafkaStub;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import garden.*;
+import garden.GardenWatcher;
+import garden.GardenWatcherInBatch;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,10 +14,7 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.time.Duration;
-import java.util.List;
-import java.util.Map.Entry;
 
-import static java.util.stream.Collectors.toList;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(classes = AppConfig.class)
@@ -24,7 +22,7 @@ import static org.assertj.core.api.Assertions.assertThat;
         SpringExtension.class,
         KafkaExtension.class
 })
-public class GardenTest {
+public class ConsumersTest {
 
     @Autowired
     ApplicationContext context;
@@ -35,8 +33,6 @@ public class GardenTest {
     @Autowired
     GardenWatcherInBatch gardenWatcherInBatch;
 
-    @Autowired
-    GardenRepo gardenRepo;
 
     @Autowired
     ObjectMapper objectMapper;
@@ -72,36 +68,6 @@ public class GardenTest {
 
         // Cleanup
         kafkaStub.unload("stubs/Garden/batchConsumerOfGardenChanges_shouldBeActive/kafka");
-    }
-
-    @Test
-    void producerOfGardenMetrics_shouldBeActive(KafkaStub kafkaStub) throws Exception {
-        // Given
-        kafkaStub.load("stubs/Garden/producerOfGardenMetrics_shouldBeActive/kafka");
-
-        // When
-        List<Fruit> expectedFruits = List.of(
-                gardenRepo.plant(new Seed("Apple")),
-                gardenRepo.plant(new Seed("Banana")),
-                gardenRepo.plant(new Seed("Orange")),
-                gardenRepo.plant(new Seed("Kiwi")),
-                gardenRepo.plant(new Seed("Watermelon"))
-        );
-
-        List<Entry<Seed, Fruit>> gardenEvents = kafkaStub.consumeMessages(
-                gardenRepo.getGardenMetricTopic(), expectedFruits.size(),
-                Seed.class, Fruit.class, 10000
-        );
-
-        List<Fruit> actualFruits = gardenEvents.stream()
-                .map(Entry::getValue)
-                .collect(toList());
-
-        // Then
-        assertThat(actualFruits).containsExactlyInAnyOrderElementsOf(expectedFruits);
-
-        // Cleanup
-        kafkaStub.unload("stubs/Garden/producerOfGardenMetrics_shouldBeActive/kafka");
     }
 
 }
