+++ brave-kafka/src/test/java/gardentest/StreamsTest.java
@@ -1,0 +1,20 @@
+package gardentest;
+
+import brave.extension.KafkaExtension;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+@SpringBootTest(classes = AppConfig.class)
+@ExtendWith({
+        SpringExtension.class,
+        KafkaExtension.class
+})
+public class StreamsTest {
+
+    @Test
+    void streamOfWater_shouldBeActive() {
+
+    }
+}
